refactor(speculative): decouple P-EAGLE from the EAGLE-3 code path
Move the P-EAGLE (parallel-drafting) support code out of the shared EAGLE-3
modules into dedicated files. Behavior-preserving: every line is relocated
verbatim and the eagle3/peagle unit tests stay green.

- core.py: PEagleTrainerModule, _PeaglePlan and the P-EAGLE _kl_div_loss move
  to peagle_trainer.py (the shared Eagle3StepMetrics stays in core.py).
- draft_llama.py: the forward_peagle paths, the COD flex-attention helpers and
  the mask_hidden handling move to peagle_draft.py as mixins the EAGLE-3 draft
  classes inherit.
- train_eagle3.py: the P-EAGLE setup body and the per-segment step methods move
  to PeagleRecipeMixin in peagle_recipe.py.

Two pieces stay in the shared files on purpose:
- the dispatch (`if parallel_drafting` in setup, `if self._peagle_partitioned`
  in the train loop) is control flow, not P-EAGLE logic; removing it would
  change behavior.
- Eagle3LlamaPeagleLayer stays in draft_llama.py because it is built from the
  shared Eagle3Llama attention/MLP blocks; moving it out would create a circular
  import.
