+++ nemo_automodel/components/speculative/eagle/__init__.py
@@ -25,11 +25,12 @@ Import them directly from those modules.
 """
 
 from nemo_automodel.components.speculative.eagle.backend import Eagle3TargetBackend
-from nemo_automodel.components.speculative.eagle.core import Eagle3TrainerModule, PEagleTrainerModule
+from nemo_automodel.components.speculative.eagle.core import Eagle3TrainerModule
 from nemo_automodel.components.speculative.eagle.core_v12 import EagleTrainerModule
 from nemo_automodel.components.speculative.eagle.draft_gpt_oss import GptOssEagle3DraftModel
 from nemo_automodel.components.speculative.eagle.draft_llama import LlamaEagle3DraftModel
 from nemo_automodel.components.speculative.eagle.draft_llama_v12 import LlamaEagleDraftModel
+from nemo_automodel.components.speculative.eagle.peagle_trainer import PEagleTrainerModule
 from nemo_automodel.components.speculative.eagle.registry import (
     EAGLE1_DRAFT_REGISTRY,
     EAGLE3_DRAFT_REGISTRY,
